add the possibility to add an action to the screencast.
modified files:
template/actions.html : give the possibility to choose the type of action
thanks to the templateHelpers on views/actions.js.
views/actions.js :
create the appropriate model and add it on the collection with
the fuction onCreateAction.
at the creation of an action, scroll on it thanks to the function scrollOnAction.